Notes on manifolds and calculus advances on linear algebra
The section on manifolds are expanded. Added a small section
on set theory I will continue to add stuff from lee's book
However the actual topic is briefly covered up to differentiable
manifolds.

For calculus I added the proof for derivative of f(x) * g(x).
I should expand it to multivariate calculus to make a passage
to manifolds.

Advanced mostly on linear algebra:

Covered almost entirely the first chapter of Serge.
Projection of vectors to each other, parametric lines
and planes are covered in brief detail. The length
of arbitrary points to line is yet to be finished.
But I might pass directly to matrices and line
equations.